pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemini_terminal"
version = "0.1.0"
description = "A terminal-based UI for Google's Gemini AI with MCP tool integration"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
    "textual>=0.40.0",
    "google-generativeai>=0.3.0",
    "pyyaml>=6.0",
    "mcp>=0.1.0",
]

[project.scripts]
gemini-terminal = "gemini_terminal:run"

[tool.hatch.build.targets.wheel]
packages = ["gemini_terminal"]

[tool.ruff]
line-length = 100
target-version = "py310"
